update openocd config and js logic
