+++ app/balena-fin-v1.1.cfg
@@ -5,14 +5,11 @@ sysfsgpio_tdi_num 11
 sysfsgpio_tdo_num 25
 sysfsgpio_tck_num 10
 sysfsgpio_tms_num 9
-sysfsgpio_swdio_num 32
-sysfsgpio_swclk_num 33
+sysfsgpio_swdio_num 33
+sysfsgpio_swclk_num 32
 sysfsgpio_srst_num 510
 
 transport select swd
 
 # Set parameters for BGM111
-set CHIPNAME efm32
-source bgm111.cfg
-
-reset_config srst_nogate
+source /usr/src/app/openocd-code/tcl/target/efm32.cfg
