--- app/flash.sh
@@ -1,1 +1,0 @@
-openocd -f balena-fin-v1.1.cfg
