+++ app/start.sh
@@ -1,14 +1,1 @@
-echo "toggling co-processor mux to flash mode..."
-if [[ -e /sys/class/gpio/gpio41 ]]; then
-  echo 41 > /sys/class/gpio/export && echo "out" > /sys/class/gpio/gpio41/direction && echo 1 > /sys/class/gpio/gpio41/value
-fi
-sleep 3
-chmod +x flash.sh
-echo "Opening screen terminal"
-screen -dmS ftdi_program  "./flash.sh"
-echo "Opened screen terminal"
-sleep 5
-{ sleep 5; echo "reset"; echo "program firmware/bootloader.s37"; echo "program firmware/firmata.hex"; echo "reset run"; sleep 10; } | telnet localhost 4444
-sleep 5
-echo "RUN $ node main.js, to start firmata checks"
-tail -f /dev/null
+
